fix: mobile sticky header revert of the revert (#10079)
