+++ src/pages/Dashboard/components/DashboardHeader/DashboardHeader.tsx
@@ -1,15 +1,22 @@
 import type { ResponsiveValue } from '@chakra-ui/react'
-import { Container, Flex, Stack, useColorModeValue } from '@chakra-ui/react'
+import { Container, Flex, Stack, useColorModeValue, useDisclosure } from '@chakra-ui/react'
 import type { Property } from 'csstype'
+import { useScroll } from 'framer-motion'
 import type { JSX } from 'react'
-import { memo, useEffect, useLayoutEffect, useMemo, useRef } from 'react'
+import { memo, useCallback, useEffect, useLayoutEffect, useMemo, useRef, useState } from 'react'
 import { useLocation } from 'react-router-dom'
 
 import { DashboardTab } from '../DashboardTab'
+import { DashboardDrawer } from './DashboardDrawer'
 import { DashboardHeaderTop } from './DashboardHeaderTop'
 import { DashboardHeaderWrapper } from './DashboardHeaderWrapper'
+import { MobileUserHeader } from './MobileUserHeader'
 
 import { Display } from '@/components/Display'
+import { GlobalSearchModal } from '@/components/Layout/Header/GlobalSearch/GlobalSearchModal'
+import { MobileWalletDialog } from '@/components/MobileWalletDialog/MobileWalletDialog'
+import { useModal } from '@/hooks/useModal/useModal'
+import { isMobile as isMobileApp } from '@/lib/globals'
 
 const paddingTop = {
   base: 'calc(env(safe-area-inset-top) + var(--safe-area-inset-top))',
@@ -17,6 +24,9 @@ const paddingTop = {
 }
 const marginTop = { base: 0, md: '-4.5rem' }
 const borderBottomWidth = { base: 0, md: 1 }
+
+// If we set this to 0, the transparent background will cause some weird flickering when scrolling back to 0 or opening the drawer
+const TRIGGER_BACKGROUND_HEIGHT_Y = 2
 
 export type TabItem = {
   label: string
@@ -39,6 +49,27 @@ export const DashboardHeader = memo(() => {
   const location = useLocation()
   const activeRef = useRef<HTMLButtonElement | null>(null)
   const containerRef = useRef<HTMLDivElement | null>(null)
+
+  const qrCode = useModal('qrCode')
+  const [y, setY] = useState(0)
+  const { scrollY } = useScroll()
+
+  useEffect(() => {
+    return scrollY.onChange(() => setY(scrollY.get()))
+  }, [scrollY])
+
+  const handleQrCodeClick = useCallback(() => {
+    qrCode.open({})
+  }, [qrCode])
+
+  const {
+    isOpen: isSearchOpen,
+    onClose: onSearchClose,
+    onOpen: onSearchOpen,
+    onToggle: onSearchToggle,
+  } = useDisclosure()
+
+  const { isOpen, onClose, onOpen } = useDisclosure()
 
   const borderColor = useColorModeValue('gray.100', 'whiteAlpha.200')
 
@@ -133,18 +164,51 @@ export const DashboardHeader = memo(() => {
     }
   }, [])
 
+  const mobileDrawer = useMemo(() => {
+    if (isMobileApp) return <MobileWalletDialog isOpen={isOpen} onClose={onClose} />
+    return <DashboardDrawer isOpen={isOpen} onClose={onClose} />
+  }, [isOpen, onClose])
+
   return (
-    <DashboardHeaderWrapper>
-      <Stack
-        spacing={0}
-        borderColor='border.base'
-        borderBottomWidth={borderBottomWidth}
-        pt={paddingTop}
-        mt={marginTop}
-      >
-        <DashboardHeaderTop />
-        <Display.Desktop>{tabs}</Display.Desktop>
-      </Stack>
-    </DashboardHeaderWrapper>
+    <>
+      <Display.Mobile>
+        <Container
+          px={4}
+          position='fixed'
+          top='0'
+          pt='calc(env(safe-area-inset-top) + var(--safe-area-inset-top) + var(--chakra-space-4))'
+          zIndex='banner'
+          bg={y > TRIGGER_BACKGROUND_HEIGHT_Y ? 'background.surface.base' : 'transparent'}
+          pb={4}
+          maxWidth='100%'
+        >
+          <MobileUserHeader
+            onSearchOpen={onSearchOpen}
+            handleQrCodeClick={handleQrCodeClick}
+            onOpen={onOpen}
+          />
+        </Container>
+
+        <GlobalSearchModal
+          isOpen={isSearchOpen}
+          onClose={onSearchClose}
+          onOpen={onSearchOpen}
+          onToggle={onSearchToggle}
+        />
+        {mobileDrawer}
+      </Display.Mobile>
+      <DashboardHeaderWrapper>
+        <Stack
+          spacing={0}
+          borderColor='border.base'
+          borderBottomWidth={borderBottomWidth}
+          pt={paddingTop}
+          mt={marginTop}
+        >
+          <DashboardHeaderTop />
+          <Display.Desktop>{tabs}</Display.Desktop>
+        </Stack>
+      </DashboardHeaderWrapper>
+    </>
   )
 })
